Converted the unstructured hex mesh example to using the mesh API (as it currently exists).
Uses UnorderedSet to encode the nodes and zones of the mesh.
Uses StaticVariableRelation to encode the incidence relations from nodes to zones and from zones to nodes.
Note: since the latter is a constant relation, we will change its implementation once constant relations are available.
Similarly, while we are building the node->zone relation, we should encode this as a dynamic variable relation.