ENH: minor refactoring
Logger now holds a collection of log streams. Messages
regardless of level are pushed to all of the streams. Further,
to make the API more compact, removed the enable/disable methods
with a single "setLoggingLevel" method. Last, changed ownership
semantics of log stream object. Although, the application allocates
the LogStream object the ownership of the pointer is propagated
to the Logger instance that holds it. Examples have also been
modified to reflect these changes.