ENH: Added MPI support for macros
Teach make_library() and make_executable() about MPI.
Specifically, when "WITH_MPI" is passed as an argument
to either macro,  the macro will include the MPI include
path and the -DUSE_MPI, as well as, other MPI compile flags
to the compiler command. For make_executable(), the MPI
C libraries are also linked with the given target.

This required changing make_library to accept token based
cmake arguments. Thus, modified the `meshapi` and `sidre`
CMakeLists.txt to reflect this change.

Also, added a new macro, 'add_target_definitions', which is
similar to cmake's native 'add_definitions', but, allows
for more fine-grained control of pre-processor definitions
on a per-target basis.

Last, I tried to improve the cmake documentation for the
macros.

Conflicts:
	src/cmake/CMakeBasics.cmake