ENH: Bind logstreams to levels
This commit addresses JIRA issue ATK-174.

Support is added for binding logstream to levels.
For example, a code may specify that all errors
and warnings go to a particular file, while a more
concise form of all the message is printed to the
console. Also, added an example, cpplogger_example to
illustrate this.