ENH: replace static API with unbound methods
This commit addresses ATK-175 and ATK-209.

For clarity, the static API is moved out of the Logger class, with
the exception of initialize(), getInstance() and finalize() methods.
Instead, the logger API is implemented using unbound methods within
the slic namespace. To avoid confusion, the unbound methods have the
same name as the equivalent Logger instance methods.

Further, slic::isInitialized() was added to check if the logging
environment has been initialized. If the logging environment is
not initialized, slic::logMessage() now returns
immediately (resolves ATK-209).

Last, the examples and macros have been updated to reflect
the latest changes.