Improves the cmake make_executable macro to allow developers to choose the name of the target executable. Resolves ATK-224.
Also improves and tests a previous commit providing an option to enable openmp.
This commit adds a small openmp exe which is built only when openmp is enabled and which renames the target exe. Tests ATK-220 and ATK-224.

Also fixes a test in the common component (common_utilities).  This test was no longer being built after the name was changes from cxx to cpp.
Once I reenabled the test, I had to fix the test.