Meshapi port with Lulesh checked against MPI.  ATK-212. (Note: OpenMP correctness was checked in an earlier commit, and I'll check both again after all other edits before resolving ATK-212).
Bugfix: The face adjacencies (m_{xi,eta,zeta}_{m,p}) are relations from the elements to the 'extended elements' (elements plus ghost elements on the surface of the domain ).
Bugfix: Cycle and error checker in Lulesh depends on the overall number of elements on an edge of the problem, not just those on an edge of the domain.

Misc:
* Lulesh now builds with MPI using the cmake make_executable macro, using a custom name.
* Looked into logic of MPI communications a bit.  This can be compacted and simplified signficantly using a small struct that keeps track of ranks and offsets to neighbor data.