Add flags to tell fortran where to search for modules.
Xlf's -qmoddir is added by CMake to define where module files are
written to. For many compilers this also implies that the directory
should also be used to look for module files. Not so with xlf. Need to
tell xlf where to look for modules.

Switch to the 2003 compiler since we're using features from that standard.
ATK-219