ENH: hook for filtering out duplicate messages
Added optional flag to all logMessage() methods that indicates whether
duplicate messages, coming from multiple ranks, need to be filtered
out. Note, it is the responsibility of the underlying stream to implement
the logic for identifying and combining duplicate messages. This commit
simply ensures that the interface is capable of propagating this request
from the application to the underlying LogSteam instance. By default,
this parameter is set to false.