shroud: Add option F_string_len_trim This controls the action of function string_to_buffer_and_len
For each function with a ``std::string`` argument, create another C
function which accepts a buffer and length.  The C wrapper will call
the ``std::string`` constructor, instead of the Fortran wrapper
creating a ``NULL`` terminated string using ``trim``.  This avoids
copying the string in the Fortran wrapper.

Function isNameValid, which as a C_code option, was generating an
error because the string length argument was unused since the C_code
block was not construction a std::string, but only looking for NULL.
There is no way (yet) to set options on a generated routine, like
setting C_code to use both the buffer and length arguments.