Created new script to automatically prepend copyright statement to source files. This can be used for ATK-246.
For each [c/c++ header and source] file in a provided directory, the script checks if the top lines contain the copyright notice.
If not, it adds the copyright notice and saves the file.

Some additional features that might be nice:
* specify the valid file types
* allow recursive traversal of the file system
* allow test runs that indicate which files will be modified without actually changing anything

The script was based on a script from P. Sinha given to me by G. Zagaris.