Bugfix: Added an option to enable/disable google benchmark functionality. Related to ATK-351.
There was a conflict between the build system for the (third party) google benchmark library and the toolkit on non-gnu compilers.
The problem appears to be related to the use of C++11 and flags that gbenchmark sets vs. those that we set.

Since the libraries work with our gnu compilers on LC, but not with our clang or intel compilers,
I have disabled this feature by default, and enabled it only on gcc (in the corresponding host_files).

This is controlled by the ENABLE_BENCHMARK option in cmake.
I will set up an issue in jira to fix this for the other compilers.