Meshtk extent now correctly handles cell indexes.
The cell indexes is a bit of a hack and should be refactored.
Specifically, we probably want to simplify the logic so that cell extent
and node extent are different instances of an extent.
StructuredMesh will then handle the interaction between the two.

Also added print functions to quest's Point and BoundingBox classes

Conflicts:
	src/components/quest/examples/quest_driver.cpp
	src/components/quest/src/geom/BoundingBox.hpp