ENH: Segment primitive & associated operations
This commit adds a Segment class, which can be used to represent
a finite oriented line, defined by a source and target point, mainly
in 2-D and 3-D. Further, implemented associated geometric operators
to query the orientation and distance of a point with respect to
a given segment. Last, additional tests were added in test_quest_orient
and test_quest_squared_distance respectively to verify basic functionality
of the new primitives.