Adding tests to quest's Vector class. ATK-457.
Fixed a bug in Vector's normalize() function -- normalizing the zero vector had a divide by zero problem.
The zero vector is now converted to the unit vector (1,0,0,..) when normalized.
Also improved construction of vector to give a similar interface as its parent class (Point).

TODO: Still have to add tests for the arithmetic operators and outer and inner products.