Refactored quest's point and vector in terms of NumericArray. ATK-460.
NumericArray encapsulates a statically sized array and adds
a set of component-wise arithmetic operations (+,-,*,/,min,max,...).
The vector and point classes are now implemented with a NumericArray
(no inheritance), and expose an array() method to access this,
as well as a data() method to access the underlying array.

Arithmetic operations on Points and Vectors can now be implemented
in terms of the corresponding operation on their NumericArray member.

Also removed the normalize() function from quest::Vector, which modified
the vector and had problematic semantics for zero vectors.
I replaced this with a const unitVector() function which
returns a unit vector (and the vector (1,0,0,..,0) for the
unit vector.

TODO: We need to be able to let users decide the EPS threshold for
determining when a vectors is (fuzzy)-equal to the zero vector.
This is currently hard-coded into the unitVector() function.