ENH: Add 2D Ray-Segment intersection operation
This commit introduces a 2D ray-segment intersection operator.
The call to quest::intersection() returns true if an intersection
exists and the point of intersection is returned through the
argument list. In addition, a test to ensure correctness of the
basic functionality is added. Last, as with the SquaredDistance
and Orientation operators, the method can be overloaded in the future
to support additional geometric primitive types, e.g., Ray-Ray,
Ray-Triangle intersections, etc.