ENH: Nullify requirement for normalized vector
This commit nullifies the requirement for a normalized
vector to construct a Ray object. Instead, the user-supplied
is normalized by calling unitVector() in the constructor.
Updated, the assertion inthe construct to reflect this change
and modified the documentation accordingly. Also, improved the
documentation for Ray::at( t ) function which was only partially
complete.