ENH: implement function to bisect a boundingbox
This commit implements a function to bisect a bounding box
into two sub-boxes along a given dimension. The caller can
speficy the dimension at which to bisect the bounding box
via an optional argument to the function. If a dimension is
not supplied, the code will bisect the bounding box along
the longest dimension.

Additional support methods, e.g., to query the longest dimension,
and setters needed for this functionality were also added in this
commit.

Last, new unit tests were added to test the new functionality.