Created a wrapper for the fixed-width types in common and removed dependency on boost. ATK-538.
Details: Fixed width types can be a bit tricky.  They were defined in C99, but are not
part of the C++ standard until C++11.  We are using a snippet borrrowed from http://snipplr.com/view/18199/stdinth/
when C++11 is not available, and hiding these implementation details in a wrapper file.