Adds mesh vertices to an InOutOctree for accelerating point containment queries.  ATK-440.
This commit creates a InOutOctree as a derived class from SpatialOctree on the way towards
our point containment acceleration structure..

In this first step, we insert the vertices from the triangle surface mesh into the octree
under the condition that a leaf node from the octree can index at most one vertex from the mesh.

Once the vertices are inserted, we reindex the mesh vertices according to the octree's ordering.
This ensures that all triangles incident on a common mesh vertex have the same id for this vertex.
(Note: The input STL file does not have such a consistent index for the mesh vertices).

Misc:
* Refactor quest's octree and octree_driver
* Preliminary profiling showed an inefficiency with the function in SpatialOctree that quantized points onto a regular grid cell.
  This revealed that we can simply cast the grid cell to an integer rather than calling the floor function.
  This gave a ~25% improvement in runtime for inserting the mesh vertices.
* Modified constraints on OctreeBase's templated LeafData type.
  Specifically, a valid LeafData must be default constructible, and provide an isLeaf() function as well as a setInternal() function.
  This is indicated in the class comments for OctreeBase.
  Note: This is subject to change after we convert the implementation to use more of the SLAM components.