Bugfix for clang compiler error on 3D Morton indexing. ATK-571.
It turns out that clang's optimizer (for -O2 and above)
has a bug on the MortonBase::expandBits() function.

I found a temporary workaround --- marking the loop variable as 'volatile' disables the erroneous optimizations.
The problem is now fixed, and I also created a working reproducer for the clang optimization bug (sent to LC),
which is being run under our test suite.  If/When the problem is resolved in clang, the test should fail,
which will let us know to we can remove the workaround.