BUGFIX: fix BoundingBox default constructor
The default constructor was using numeric_limits to initialize
min/max of the bounding box, s.t. the box is invalid. Specifically,
min was set to numeric_limits< T >::max() and max was set to
numeric_limits< T >::min(), where T is the underlying datatype,
e.g., double, float, int etc.

However, numeric_limits< T >::min(), when T is a floating-point,
returns the minimum *positive* normalized value. This was not
the intended behavior and is different than the behavior of
min() when T is an integral type (i.e., int, long, etc.).

This led to errant construction of bounding boxes when constructed
incrementaly via the addPoint() method. This commit resolves the
issue by properly initializing m_max of the BoundingBox to
(-1)*numeric_limits< T >::max().