Bugfix: The bounds for the default constructor in quest::BoundingBox were wrong for float and double types.
George Zagaris found a bug in our code.  std::numeric_limits<double>::min() is actually the smallest
positive double value, but our code assumed that it was the smallest representable negative number.
This was fixed in CXX11 with the introduction of a new function:  numeric_limits::lowest().

This bugfix introduces a type trait to use the CXX11 version when posible with a fallback to valid pre-CXX11 code.