Use function names createViewScalar and createViewString instead of overloading createView. createView(std::string, char *) is ambiguous.  It could mean the user wants to create an external view or a string view.
  template<typename ScalarType>
  DataView * createView( const std::string& name, ScalarType value)
Created a compile error in the lulesh example for createView(string,
Domain *).  The template was being used instead of the void * overload
to create an external view.  The call to setScalar triggered the error
since there was no type trait and it could not be converted to a Sidre
type.