ENH: Add non-const reference accessors to triangle
This method adds non-const reference access methods to the triangle
vertices to simplify construction of a triangle instance and in an
attempt to minimize the number of copies of quest::Point instances
for each triangle vertex. Also, made the default constructor public.
Application code can now create a Triangle instance, which results
to a degenerate triangle and use the non-const reference access methods
to set the triangle vertices.