ENH: parallel STL reader
Implements a parallel STL reader inheritting from the serial
STL reader. This required making the destructor and the read()
method of the STL reader virtual since it is now a base class.
Rank 0 reads the data from the STL file. Then the data is
copied to all other ranks through a broadcast.