Added containment queries to quest's interface and exposed to fortran.  ATK-676.
Quest interface changes:
* quest_initialize now has an additional parameter -- bool requiresDistance.  If false, we use the InOutOctree, if true, we use the SignedDistance.
* The return value of quest::inside() now better matches the intuitive meaning. The function returns 1 if the point is inside, and 0 otherwise.

The interface examples (C++ and Fortran) now exercise both types of queries -- signed distance and containment queries.

This commit also simplifies and encapsulates the logic of the quest interface functionality.
The code still assumes a single 3D instance of the acceleration structures, but it will be much easier to set it up for multiple such structures now.

Notes:
* We will likely need to modify the initialize function in the future to take a struct with the appropriate initialization parameters for quest.
  In particular, the various structures have specific parameters (e.g. bucket size and depth for SignedDistance) that are not necessary for the
  other data structures.  Also, we might want to set parameters about the mesh, e.g. if it is already read in, and how many parts it has, etc...