Bugfix for Quest::InOutOctree -- withinGrayBlock() point containment. ATK-715.
Original approach to determining containment of points within gray blocks was not properly handling coplanar triangles,
or situations where the local geometry cannot be classified as strictly concave or convex.
For the former, we now use the normalized triangle normals.
For the latter, we need to find the distance to the closest triangle and handle situations where the closest point is on a triangle edge or vertex.

The 'checkValid' function now verifies that Inside blocks of the octree are not face-adjacent to Outside blocks of the octree.

A good problematic region is on the 'plane_simp.stl' model around the bounding box with min point: (68.326,740.706,187.349) and max point: (68.5329,740.923,187.407).