uberenv: add new install helper
added initial python script that creates a timestamped tpl
builds for all compilers we care about on chaos 5 platforms.

note: I am not changing the existing lc install scripts.

It appears they were changed to support bamboo, in a way that
broke  their use outside of bamboo.

The primary goal of these scripts is to work OUTSIDE of bamboo,
so that anyone on the toolkit team can wield them for TPL installs.