uberenv: disable hdf5 fortran support
We are not using hdf5's fortran interface and trying to build it
chaos systems with clang (which do not have a companion fortran
compiler) raises an error and stops our spack tpl install process.

The spack hdf5 package enables fortran support by default,
and then throws an error if a fortran compiler isn't found.

To avoid this error, we turn off the hdf5 fortran variant.