blt: fixes that allow boost to be used in slam
We are using boost as header only, I changed blt_register_lib
so if no libraries are passed we set a special value to differentiate
this from a generic external dependency.

(without this change, we end up trying to link "-lboost")

This is a temporary fix. I am starting to worry (again) about BTL support
for transitive dependencies. We can use imported targets to solve this for
libraries, but I don't know if we can (or want to ) for headers.

Note: I am 99% convinced the trunk is only working on LC b/c it is picking
up system boost headers. This is bad, and when fixing this it appears
boost is needed for the public interface for slam, we need to resolve this.
Even if we use boost internally, we don't want client code to depend on it.