ENH: use pseudo-normal for sign of signed distance
Compute and use a pseudo-normal to more accurately calculate the
sign of the signed distance. Further, added code to calculate
the bounding box of the object (i.e., surface mesh). Query points
outside this bounding box are known to be outside and therefore
the sign does not need to be computed. Filtering those points out
helps resolve artifacts on the corners and also speeds up the code
by avoiding unnescessary sign computations.

Resolves ATK-718.