Renamed the fmt/detail directory to fmt/fmt. ATK-580.
This allows us to directly use the code from the thirdparty fmt library without any modifications.

Note: The original code is in a separate directory since it needs to be configured with some compiler definitions to be used by the ASC toolkit.
See setup code in src/thirdparty/fmt/fmt.hpp