Bugfix for failing tests when building in a different directory.
Some tests in common and slam hardcoded a path to mesh files, and assumed the code would be build in a sibling directory to the src directory.
We now pass a path in as a command line argument using the blt add test macro.

Note: We should revisit this once we have set up a data repository.

Misc: Using fmt library in some slam examples.