Changes to Fortran API for save/load
Add function view%set_array_data_ptr(array)
This is similar to view%set_external_data_ptr except it takes a
Fortran array instead of a type(C_PTR) making it more natural to use.
    call view%set_external_data_ptr(c_loc(array))
    call view%set_array_data_ptr(array)
In addition, since set_array_data_ptr is taking an array the type and
shape information is available.  In the future it could be used to
check consistency between the array and the view

Add function grp%set_array_data_ptr("name", array)
This is similar to the C++ code
    grp->getView("name")->setExternalDataPtr(array)
In Fortran this would look like
    view = grp%get_view("name")
    call view%set_array_data_ptr(array)

Increase number of dimensions of Fortran generic functions from 3 to
4. The standard allows 7 and the 2008 standard raised it to 15.  I'm
only using 4 as the amount actually needed for Miranda and to minimize
the amount of generated code for now.

Wrap some additional function in Fortran.
  DataStore  save, load, loadExternalData
  DataGroup  isEquivalentTo
  DataView   isDescribed, isEmpty

Add additional Fortran tests to save and load a DataStore.

Add script cpp2fortran to help convert cpp tests to Fortran.
It does ~70% of conversion using regular expressions.