ENH: setup for development script
Adds script to setup the cloned repo for development. Currently,
the script ensure that git is configured correctly, i.e., there
is an associated e-mail and name etc. Also, the script will ask
to setup useful aliases to streamline development, enable git
colors, setup editor etc.

The intent of this script is to simplify setup for new developers
and/or the migration to a new workstation where git is not setup
for the user.