Add cpp_local_var to Shroud
This variable will create a local variable for an argument instead of
passing the argument directly to the function.  It uses c_to_cpp to
assign the value.  Often this value can be very long so a local
variable is nicer.  May help with debugging too since an intermediate
variable is created.