bugfix: fix crash with DataView::getData()
fixed bug with return for getData(), getScalar() and getArray()
for uninited views.

You can't call Node::value() on a temp Node instance since
it uses a pointer to the source node. This is why these
calls in DataView crashed instead of throwing exceptions.

The solution is to call m_node.value(), which will provide
a default value (0 for non-pointer results and NULL for
pointer results).

Made datastore rewire conduit warning and info handlers to slic.

Added unit test for getData(), getScalar(), and getArray()
for the case when the view is uninited.