Restore EANBLE_CXX11 option
This is necessary for compilers like intel which say they have CXX11 support
but are not configured properly to use a CXX11 stdlib. This fixes build errors like:
  /usr/include/c++/4.4.7/c++0x_warning.h(31): error: #error directive:
  This file requires compiler and library support for the upcoming ISO C++ standard, C++0x.
  This support is currently experimental, and must be enabled with the
  -std=c++0x or -std=gnu++0x compiler options.

Update bamboo scripts to echo arguments to config-build.py.
This can be to cut-and-paste from the bamboo log to reproduce the config line.