Add getPath() and getPathName() to Group and View.
This patch adds methods getPath() and getPathName() to DataGroup and
DataView classes, joining getName().  For a DataStore containing DataGroup
tree a/b/c/d/e, with group d owning a view v, the following are expected:

  d->getName()     d
  d->getPath()     a/b/c
  d->getPathName() a/b/c/d
  v->getName()     v
  v->getPath()     a/b/c/d
  v->getPathName() a/b/c/d/v

The patch adds unit tests of DataGroup and DataView unit tests to verify
this behavior.

To support the implementation, and for consistency, the DataStore's root
group now returns an empty string for its name, path, and path name.
Additionally, the root's parent group is set to be itself.