Adds MortonOctreeLevel to Quest's octree.
MortonOctreeLevel uses a MortonIndex (e.g. std::size_t) as the key
rather than an integer grid point (e.g. Point<int,DIM>) for
the block data of an octree level.

Note: We can use this for levels where level*DIM <= 64.

The next step is to specialize the class so we can use a 16 or 32 bit
int as the key, rather than a 64 bit int, when level*DIM < {16,32}.