Specialized MortonOctreeLevel on the type necessary to index all points in a level.
Added a template parameter to Mortonizer and MortonBase for the type of the Morton index.
We are now using fewer iterations to mortonize and demortonize based on the desired output type.
We are also using a smaller type as the key in the MortonOctreeLevel hashmap from
octree blocks to associated data.