Simplified data access in Quest's OctreeLevel iterator.
Previously, we needed a data() member function for mutable access the data (of type BlockDataType)
pointed to by the iterator.  By fixing the iterator's template parameter types, we can now get either the
reference or const reference in the dereference() function (depending on
whether it is a mutable or const iterator), which removes the need for the data() function.
(E.g.   BlockDataType& bd = *it;  // rhs was previously it.data() )

Modified the data access in InOutOctree.

Misc: Updated debugging code (to query and dump an InOutOctree) to reflect
changes to the iterator API.  This code is typically disabled via a compiler #ifdef.