Refactored the derived OctreeLevel classes into SparseOctreeLevel and DenseOctreeLevel.
SparseOctreeLevel uses a map from a point (represented as either a MortonIndex
(unsigned int of specified bit width) or an integer grid point.

DenseOctreeLevel uses an array with one data item for each block in the level.
For efficiency, the array is ordered by Morton index, and the blocks are stored in broods.