ENH: add 'MINT_' prefix to mint enums
To avoid conflicts with predefined types in other codes both
the 'MeshType' and 'CellType' enumerators are now prefixed
with 'MINT_'. Further, took the enum definition out of the
'mint::' namespace to make use of the code less verbose and
since enums are now namespaced in the C-sense with the 'MINT_'
prefix. Last, updated rest of the code to reflect these changes.