ENH: robustify and speed up signed distance query.
This commit effectively reimplements the signed distance
algorithm. The algorihm is now broken into 3 main parts,
described in more detail below:

(1) Candidate Search: Given the set of BVH bins required
to satisfy the query, the candidate search, first sorts
the bounding boxes of the surface elements by their
distance to the point. The reduce set of candidate surface
elements is then obtained by looking at the surface element
with the closest bounding box and neighboring surface elements
within that vicinity, i.e., all surface elements whose
bounding box intersects with the closest bounding box.

(2) Closest Point/Min Distance: Next, the list of candidate
surface elements is processed and the closest point is
determined using the closet point primitive from a point
to a surface element.

(3) Sign Determination: Last, the sign is now more accurately
determined by using a more appropriate pseudo-normal when needed.
If the query point is outside the bounding box of the user-supplied
surface mesh, the sign is known and no pseudo-normal is needed.

Otherwise, the pseudo-normal, denoted by N, is calculated based
on the location of the closest point, C, as follows:
* N is simply the surface element normal if C is in the surface element
* N is given by the sum of surface element normals if C is on an edge
* N is given by the angle-weighted pseudo-normal if C is on a node

In addition, the quest-driver was updated to facilitate debugging, etc.