Restructured code for readability.
Now the only things in Intersection.hpp in the quest namespace are different
overloadings of intersect().  All the worker functions are put in the
anonymous namespace, and put after the definition of the intersect() they
support.  Worker functions are also declared before use, also in the anon
namespace.

It builds and passes the test suite.  I'll have to extend the test suite
before I have much confidence in it.