Changes semantics of Sidre's DataView::getVoidPtr().  ATK-855.
DataView::getVoidPtr() now returns the base pointer that was used to set up the view
(i.e. without accounting for the offset to the first element).  Previously, this function
returned a void pointer to the first data element.  For that functionality, one can still call
DataView::getData() and DataView::getArray(), which account for the offset.

This commit affects all callers of DataView::getVoidPtr().
I have updated all Sidre tests to verify the correctness of this new functionality,
and have fixed the fortran wrappers for DataView::getData(), which internally calls DataView::getVoidPtr().
(Thanks Lee Taylor for help in dealing with C pointer arithmetic in fortran).