uberenv: set SDKROOT path explicitly for OS X
The uberenv script specifies a MACOSX_DEPLOYMENT_TARGET but does
not explicitly set SDKROOT. CMake *should* compute SDKROOT
based on the value of MACOSX_DEPLOYMENT_TARGET, but, that does not
always happen correctly. In that case, CMake appears to point
SDKROOT to an older SDK, which, may, or may not exist on the
system. If the older SDK exists, there is no harm other than
using an older SDK. However, it leads to failures immediately
if the SDK does not exist. Fixed that by setting the SDKROOT
environment variable in uberenv explicitly.