ENH: granular abort behavior toggles
This commit adds the ability to control the abort behavior
on error messages and warnings on a per-logger basis. The
application code can now call `setAbortOnError( status )`
and `setAbortOnWarning( status )` respectively to prescribe
the desired abort behavior on the currently active logger.
Corresponding enable/disable methods, e.g., enableAbortOnError(),
are also added that can be used to control the abort behavior.
Furthermore, the abort behavior can be queried via calls to
"isAbortOnErrorsEnabled()" and "isAbortOnWarningsEnabled()".
By default, the code aborts only on errors, as it was the
case in earlier versions of the code.

Last, this commit moves the abort logic from the interface
layer to Logger::logMessage() since all the calls are
funnelled through that function.