Restore gcc 4.7.1 support
Restore the work around for a gcc 4.7.1 bug involving using C_LOC with
an assumed shape array.  It was removed because we figured "who's
using an old compiler like that?". Turns out that BG/Q is.  This time
the fix is applied conditionally by checking on the define
USE_C_LOC_WITH_ASSUMED_SHAPE. This defined is created by a call to the
CMake routine try_compile.

ATK-867 sidre fortran wrappers errors on bgq