Bugfix: Added missing dependency of Slam's Shocktube on boost, when ENABLE_BOOST=ON. ATK-888.
Even though the code in the ShockTube example does not depend on boost, the configuration
defines #ATK_USE_BOOST when ENABLE_BOOST=ON, and the #include for boost headers
in slam/src/OrderedSet.hpp is guarded by ATK_USE_BOOST.

So, slam's users depend on boost headers in toolkit configurations where  ENABLE_BOOST=ON.
I verified that all other examples/tests already took care of this.

Looks like the problem was actually another variant of ATK-730.