BUGFIX: fix regex for trailing space in msg hook
Fix erroneous regex in the commit message hook, which is
used to validate the format of commit messages. The regex
tried to match trailing whitespace on the subject line using
the pattern: `grep [[ \t]]*$`. This was breaking when a commit
message was ending with the letter 't', causing the hook to
erroneously reject the commit message.

An alternative way to match trailing whitespace, namely,
using the pattern `grep [[:space:]]$`, is now employed. This
appears to be a more robust soluition and resolves the issue.

This commit addresses issue ATK-897.