Add make targets to compile shroud tests.
Compile and run the run-tutorial and run-strings tests.

Remove use of common/FC.h. This header file is created by CMake to
create a macro, FC_GLOBAL, which mangles C routines so they are
callable from Fortran. There is only one routine in shroudrt.cpp which
needs this macro. So instead of recreating the macro, I added two
copies of the routine: shroud_c_loc and shroud_c_loc_. That way it can
be called from any compiler we care about (for now).