COMP: add dummy primal source to build primal
Currently, BLT does not support header only libraries well.
To circumvent the issue, this commit adds a dummy source, primal.cpp,
with an empty function. This source is compiled in to libprimal.a and
is used to trick the build system in to treating primal as a normal
library with sources.

This commit resolves in part ATK-599.