Make methods const where possible.  Improve test cases.
Also add call to floor() back in, within getBinIndex().  This is needed
to properly deal with coordinate values where x_k is between -1 and 0.
In tests, template VirtualGrid instances on int, not on the point type,
avoiding cognitive dissonance when inserting into a range of bins.
Reverse argument order in calls to EXPECT_EQ.  Test a point
(100., 100., 100) on the upper corner of a VirtualGrid.