ENH: Add Matrix class and associated operators
Add Matrix class in axom_utils. The matrix class is
used to store M x N matrices and provides a natural
way to index the coefficients of a matrix using the
row and column indices, regardless of the underlying
flat array storage (column-major).

In addition, the Matrix implementation provides
overloaded operators for:
* matrix addition and subtraction
* matrix scalar multiplication
* matrix-vector product
* matrix-matrix multiplication
* output stream operator (<<)

Special matrix operators, e.g., to compute the
matrix transpose are also provided.

Associated unit test is also implemented.

Also, updated the doxygen config file (Doxyfile.in) for
axom_utils so that it renders Latex.

This commit resolves ATK-947.