ENH: compute determinant of arbitrary size matrix
Add method to compute the determinant of arbitrary size
NxN matrices. For N <= 4, the hand-coded, pre-existing,
implementations are used. For larger matrices, the matrix
is first factorized using LU decomposition. Then the
determinant is computed as the product along the diagonal
of the L matrix, times the parity bit to correctly
determine the sign of the determinant. Corresponding
unit test is also added as part of this commit.