ENH: Add FEM framework in Mint
This commit adds an initial implmementation of a FEM
framework in Mint. This first implementation adds the
key infrastructure pieces and classes. Moreover it
adds support for Lagrangian shape functions for
the bilinear quadrilateral and linear triangle elements.

The framework provides a polymorphic FiniteElement class
that can be bound to a Finite Element basis family at runtime.
This is the main class that application code will interface
with. The FiniteElement object provides the ability to
(1) map between the reference and physical space of the
element and (2) evaluate the shape functions and jacobian
at a given point in reference space \xi.

The implementation employs a Newton-Raphson solve for the
inverse mapping, i.e., going from physical space to the
reference space, which has quadratic convergence, but,
requires a good initial guess and a positive jacobian.
This implementation uses the center of the reference element
as the initial guess to seed the Newton-Raphson iteration.

This commit resolves in part ATK-949.