New algorithm for quest::InOutOctree::withinGrayBlock(). ATK-1039.
The original algorithm had a fundamental problem in some cases.
Essentially, it tried to find the closest point to the query point
out of the local set of triangles that it indexed.  However, in some cases
(when the triangles have really obtuse angles), the closest point
of the local triangles was not the closest point on the surface, leading
to an incorrect determination of the sign.  Since we are using a flood-fill
algorithm, this small error propagated to several neighboring blocks.

The new algorithm shoots a ray from the query point to a point on a triangle
within the block, where there is no non-local information to give us the
wrong answer.

This required adding a clipping algorithm for triangles against an
axis-aligned bounding box, and an improvement in the ray-triangle
intersection algorithm to return the hit point.