ENH: reorg FEM internals & ditch cmake macro
Reorganize the FEM internal classes. Low level classes that
implement the shape functions are now prefixed by the
corresponding basis name. For example, quad_4.hpp is renamed
to lagrange_quad_4.hpp. That way, we can copy the headers in
a flat include directory and avoid name collisions in the
future. This also alleviates the need for a custom cmake
macro to copy the headers within a sub-folders from the
top include.