Bugfix in quest::InOutOctree triangle insertion and point containment test.
Triangle insertion bug:  when a vertex is near the boundary of its containing
octree block, some of its incident triangles were not being added to that block.
We were checking for this case at the leaf level, but had to extend
this check to the ancestors of these leaves.

Point containment bug: there are some rare cases where an intersecting triangle only
grazes a bounding box but we cannot find a point within the triangle that is in the bounding box.
Specifically, primal::intersect(tri,bBox) returns true, but the clipping algorithm produces
an empty polygon (e.g. primal::clip(tri,bBox).numVertices() == 0 ). In this case,
we try to clip the triangle against a slightly expanded bounding box.